# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API MAK3RS DOCUMENTATION",
        "description": "",
        "contact": {
            "email": "info@mak3rs.tk"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/communities/all": {
            "get": {
                "tags": [
                    "Community"
                ],
                "description": "Obtenemos todas las comunidades",
                "operationId": "App\\Http\\Controllers\\CommunityController::communities",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "alias": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List Communities"
                    }
                }
            }
        },
...
        "/pieces/all": {
            "get": {
                "tags": [
                    "Pieces"
                ],
                "description": "Obtenemos todas las piezas",
                "operationId": "App\\Http\\Controllers\\PiecesController::pieces",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "community": {
                                        "description": "Community uuid",
                                        "type": "string"
                                    },
                                    "alias": {
                                        "description": "Community alias",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List Pieces"
                    },
                    "422": {
                        "description": ""
                    }
                }
            }
        },
        "/communities/pieces/{alias}": {
            "get": {
                "tags": [
                    "Pieces"
                ],
                "description": "Obtenemos todas las piezas de la comunidad",
                "operationId": "App\\Http\\Controllers\\PiecesController::piecesOfCommunity",
                "responses": {
                    "200": {
                        "description": "List Pieces"
                    },
                    "422": {
                        "description": ""
                    }
                }
            }
        },
        "/pieces/create": {
            "post": {
                "tags": [
                    "Pieces"
                ],
                "description": "Creamos una pieza",
                "operationId": "App\\Http\\Controllers\\PiecesController::create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Piece"
                    },
                    "422": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/communities/piece/add-or-update": {
            "post": {
                "tags": [
                    "Stock"
                ],
                "description": "Cuando un usuario añade o actualiza una pieza a una comunidad",
                "operationId": "App\\Http\\Controllers\\StockControlController::addOrUpdatePieceStock",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid_community": {
                                        "description": "Community",
                                        "type": "string"
                                    },
                                    "uuid_piece": {
                                        "description": "Piece",
                                        "type": "string"
                                    },
                                    "units": {
                                        "description": "",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
...
        "/users/communities": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Listado de comunidades a las que pertenece el usuario",
                "operationId": "App\\Http\\Controllers\\UserController::communities",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "description": "User",
                                        "type": "string"
                                    },
                                    "alias": {
                                        "description": "",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        }
    }
}